feat(location): ajoute citySlug et countrySlug à Location et améliore l'affichage carte - Ajoute les attributs `citySlug` et `countrySlug` au BO `Location` pour enrichir les URLs. - Met à jour l'affichage des pop-ups sur la carte pour inclure le nom du lieu, la ville et le pays.

+++ workspace/springboot-app/src/main/java/fr/memoires_vives/bll/LocationService.java
@@ -1,13 +1,18 @@
 package fr.memoires_vives.bll;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import fr.memoires_vives.bo.Location;
 
 public interface LocationService {
 	Location saveLocation(Location location);
+
 	Location getById(long locationId);
+
 	List<Location> getAllLocations();
+
 	List<Location> getLocationsInSquare(double north, double south, double east, double west);
-	Location createFromCoordinates(double latitude, double longitude);
+
+	Location createFromCoordinates(BigDecimal latitude, BigDecimal longitude);
 }
